chore: bump wasmtime

+++ crates/fluvio-smartengine/src/engine/wasmtime/limiter.rs
@@ -39,9 +39,9 @@ impl ResourceLimiter for StoreResourceLimiter {
 
     fn table_growing(
         &mut self,
-        _current: u32,
-        _desired: u32,
-        _maximum: Option<u32>,
+        _current: usize,
+        _desired: usize,
+        _maximum: Option<usize>,
     ) -> anyhow::Result<bool> {
         Ok(true)
     }
